refactor: move new-item-creation code to PuzzleLayout
